refactor: re-use ProviderEncoder
