+++ keep/cli/cli.py
@@ -6,7 +6,6 @@ import sys
 import typing
 import uuid
 from collections import OrderedDict
-from dataclasses import _MISSING_TYPE
 from importlib import metadata
 
 import click
@@ -16,8 +15,7 @@ from prettytable import PrettyTable
 
 from keep.api.core.posthog import posthog_client
 from keep.functions import cyaml
-from keep.providers.models.provider_config import ProviderScope
-from keep.providers.providers_factory import ProvidersFactory
+from keep.providers.providers_factory import ProviderEncoder, ProvidersFactory
 
 load_dotenv(find_dotenv())
 
@@ -1033,16 +1031,6 @@ def provider(info: Info):
 
 @provider.command(name="build_cache", help="Output providers cache for future use")
 def build_cache():
-    class ProviderEncoder(json.JSONEncoder):
-        def default(self, o):
-            if isinstance(o, ProviderScope):
-                dct = o.__dict__
-                dct.pop("__pydantic_initialised__", None)
-                return dct
-            elif isinstance(o, _MISSING_TYPE):
-                return None
-            return o.dict()
-
     logger.info("Building providers cache")
     providers_cache = ProvidersFactory.get_all_providers(ignore_cache_file=True)
     with open("providers_cache.json", "w") as f:
